Skip aliases from contract processing
We only want to generate Dockerfiles for the canonical device type
names and not their aliases. Ex. There is a `raspberry-pi2-alpine-python` image
but not a `raspberrypi2-alpine-python` image. This removes the aliases
before building the contract to prevent the universe reproduction to
generate combinations for those aliases as well.

Change-type: minor