Generate and commit library files on finalize
Use self-hosted runners to generate and merge changes to
Dockerfiles and Libaries on finalize.

On open PR tests, the files will be generated but discarded
as the commit hash from master is used in the generation
of library files.

The above is only for testing, and to slow
the rate of PR merges to master to allow time for previous merge
actions to complete generating.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>