Commit autogenerated files directly to PR head ref
This avoids uploading of large patch files and hoping
they merge correctly on finalize.

Instead, the commits will be done upfront and included
with the PR changes.

To avoid having local copies become out-of-date during
development, it is preferred to generate these files
as part of the initial changes.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>